Fix circular dependencies in src/thing/
