+++ src/thing/asIri.ts
@@ -1,0 +1,54 @@
+/**
+ * Copyright 2021 Inrupt Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+import { resolveLocalIri } from "../datatypes";
+import { Thing, ThingLocal, ThingPersisted, UrlString } from "../interfaces";
+import { getLocalNodeName } from "../rdf.internal";
+import { internal_isThingLocal } from "./thing.internal";
+
+type IsNotThingLocal<T extends Thing> = T extends ThingLocal ? never : T;
+
+/**
+ * Get the URL to a given [[Thing]].
+ *
+ * @param thing The [[Thing]] you want to obtain the URL from.
+ * @param baseUrl If `thing` is not persisted yet, the base URL that should be used to construct this [[Thing]]'s URL.
+ */
+export function asUrl(thing: ThingLocal, baseUrl: UrlString): UrlString;
+export function asUrl<T extends ThingPersisted>(
+  thing: T & IsNotThingLocal<T>
+): UrlString;
+export function asUrl(thing: Thing, baseUrl: UrlString): UrlString;
+export function asUrl(thing: Thing, baseUrl?: UrlString): UrlString {
+  if (internal_isThingLocal(thing)) {
+    if (typeof baseUrl === "undefined") {
+      throw new Error(
+        "The URL of a Thing that has not been persisted cannot be determined without a base URL."
+      );
+    }
+    return resolveLocalIri(getLocalNodeName(thing.url), baseUrl);
+  }
+
+  return thing.url;
+}
+
+/** @hidden Alias of [[asUrl]] for those who prefer IRI terminology. */
+export const asIri = asUrl;
