+++ src/thing/thing.test.ts
@@ -30,6 +30,7 @@ import {
   removeThing,
   createThing,
   asUrl,
+  asIri,
   isThing,
   thingAsMarkdown,
   ThingExpectedError,
@@ -61,6 +62,16 @@ import { WithAcl } from "../acl/acl";
 import { mockSolidDatasetFrom } from "../resource/mock";
 import { internal_setAcl } from "../acl/acl.internal";
 import { LocalNodeIri, localNodeSkolemPrefix } from "../rdf.internal";
+
+it("exports the public API", () => {
+  expect(asUrl).toBeDefined();
+  expect(asIri).toBeDefined();
+
+  expect(ThingExpectedError).toBeDefined();
+  expect(ValidThingUrlExpectedError).toBeDefined();
+  expect(ValidValueUrlExpectedError).toBeDefined();
+  expect(ValidPropertyUrlExpectedError).toBeDefined();
+});
 
 describe("createThing", () => {
   it("automatically generates a unique name for the Thing", () => {
@@ -810,76 +821,6 @@ describe("removeThing", () => {
   });
 });
 
-describe("asIri", () => {
-  it("returns the IRI of a persisted Thing", () => {
-    const persistedThing = mockThingFrom("https://some.pod/resource#thing");
-
-    expect(asUrl(persistedThing)).toBe("https://some.pod/resource#thing");
-  });
-
-  it("returns the IRI of a local Thing relative to a given base IRI", () => {
-    const localThing: ThingLocal = {
-      type: "Subject",
-      predicates: {},
-      url: (localNodeSkolemPrefix + "some-name") as LocalNodeIri,
-    };
-
-    expect(asUrl(localThing, "https://some.pod/resource")).toBe(
-      "https://some.pod/resource#some-name"
-    );
-  });
-
-  it("accepts a Thing of which it is not known whether it is persisted yet", () => {
-    const thing = mockThingFrom("https://some.pod/resource#thing");
-
-    expect(asUrl(thing as Thing, "https://arbitrary.url")).toBe(
-      "https://some.pod/resource#thing"
-    );
-  });
-
-  it("triggers a TypeScript error when passed a ThingLocal without a base IRI", () => {
-    const localThing: ThingLocal = createThing();
-
-    // @ts-expect-error
-    expect(() => asUrl(localThing)).toThrow();
-  });
-
-  // This currently fails because a plain `Thing` always has a `url` property that is a string,
-  // and is therefore indistinguishable from a `ThingPersisted`. Not sure what the solution is yet.
-  // Meanwhile TS users won't get a build-time error if they're passing a plain `Thing`,
-  // which is annoying but not a major issue.
-  // eslint-disable-next-line jest/no-disabled-tests
-  it.skip("triggers a TypeScript error when passed a Thing without a base IRI", () => {
-    const plainThing = createThing() as Thing;
-
-    // @ts-expect<disabled because it does not work yet>-error
-    expect(() => asUrl(plainThing)).toThrow();
-  });
-
-  it("does not trigger a TypeScript error when passed a ThingPersisted without a base IRI", () => {
-    // We're only checking for the absence TypeScript errors:
-    expect.assertions(0);
-    const resolvedThing: ThingPersisted = mockThingFrom(
-      "https://some.pod/resource#thing"
-    );
-
-    // This should not error:
-    asUrl(resolvedThing);
-  });
-
-  it("throws an error when a local Thing was given without a base IRI", () => {
-    const localThing: ThingLocal = {
-      type: "Subject",
-      predicates: {},
-      url: (localNodeSkolemPrefix + "some-name") as LocalNodeIri,
-    };
-
-    expect(() => asUrl(localThing, undefined as any)).toThrow(
-      "The URL of a Thing that has not been persisted cannot be determined without a base URL."
-    );
-  });
-});
-
 describe("thingAsMarkdown", () => {
   it("returns a readable version of an empty, unsaved Thing", () => {
     const emptyThing = createThing({ name: "empty-thing" });
@@ -1041,83 +982,5 @@ describe("throwIfNotThing", () => {
       error = e;
     }
     expect(error).toBeInstanceOf(ThingExpectedError);
-  });
-});
-
-describe("ValidPropertyUrlExpectedError", () => {
-  it("logs the invalid property in its error message", () => {
-    const error = new ValidPropertyUrlExpectedError(null);
-
-    expect(error.message).toBe(
-      "Expected a valid URL to identify a property, but received: [null]."
-    );
-  });
-
-  it("logs the value of an invalid URL inside a Named Node in its error message", () => {
-    const error = new ValidPropertyUrlExpectedError(
-      DataFactory.namedNode("not-a-url")
-    );
-
-    expect(error.message).toBe(
-      "Expected a valid URL to identify a property, but received: [not-a-url]."
-    );
-  });
-
-  it("exposes the invalid property", () => {
-    const error = new ValidPropertyUrlExpectedError({ not: "a-url" });
-
-    expect(error.receivedProperty).toEqual({ not: "a-url" });
-  });
-});
-
-describe("ValidValueUrlExpectedError", () => {
-  it("logs the invalid property in its error message", () => {
-    const error = new ValidValueUrlExpectedError(null);
-
-    expect(error.message).toBe(
-      "Expected a valid URL value, but received: [null]."
-    );
-  });
-
-  it("logs the value of an invalid URL inside a Named Node in its error message", () => {
-    const error = new ValidValueUrlExpectedError(
-      DataFactory.namedNode("not-a-url")
-    );
-
-    expect(error.message).toBe(
-      "Expected a valid URL value, but received: [not-a-url]."
-    );
-  });
-
-  it("exposes the invalid property", () => {
-    const error = new ValidValueUrlExpectedError({ not: "a-url" });
-
-    expect(error.receivedValue).toEqual({ not: "a-url" });
-  });
-});
-
-describe("ValidThingUrlExpectedError", () => {
-  it("logs the invalid property in its error message", () => {
-    const error = new ValidThingUrlExpectedError(null);
-
-    expect(error.message).toBe(
-      "Expected a valid URL to identify a Thing, but received: [null]."
-    );
-  });
-
-  it("logs the value of an invalid URL inside a Named Node in its error message", () => {
-    const error = new ValidThingUrlExpectedError(
-      DataFactory.namedNode("not-a-url")
-    );
-
-    expect(error.message).toBe(
-      "Expected a valid URL to identify a Thing, but received: [not-a-url]."
-    );
-  });
-
-  it("exposes the invalid property", () => {
-    const error = new ValidThingUrlExpectedError({ not: "a-url" });
-
-    expect(error.receivedValue).toEqual({ not: "a-url" });
   });
 });
